test(compiler): Extract custom pass plugins into standalone .cpp files
